+++ src/main/java/umc/stockoneqback/friend/infra/query/FriendFindQueryRepositoryImpl.java
@@ -9,9 +9,9 @@ import umc.stockoneqback.user.infra.query.dto.QFindManager;
 
 import java.util.List;
 
+import static umc.stockoneqback.friend.domain.QFriend.friend;
 import static umc.stockoneqback.role.domain.store.QStore.store;
 import static umc.stockoneqback.user.domain.QUser.user;
-import static umc.stockoneqback.friend.domain.QFriend.friend;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
